requirements.txt
asgiref==3.5.0
autopep8==1.6.0
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.7
defusedxml==0.7.1
Django==4.0.4
django-chunked-upload==2.0.0
django-constance==2.9.0
django-dotenv==1.4.2
django-filter==21.1
django-js-asset==2.0.0
django-markitup==4.0.0
django-mptt==0.13.4
django-phonenumber-field==7.0.1
django-picklefield==3.1
djangoajax==3.3
djangorestframework==3.13.1
djlint==1.19.7
EditorConfig==0.12.3
et-xmlfile==1.1.0
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
importlib-metadata==5.0.0
install==1.3.5
jsbeautifier==1.14.7
Markdown==3.3.6
mysqlclient==2.1.0
numpy==1.22.4
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.4.2
pathspec==0.10.1
phonenumbers==8.13.2
Pillow==9.1.0
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.10.31
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
tzdata==2022.1
XlsxWriter==3.0.6
zipp==3.8.0

selenium~=4.10.0